requirements.txt
amqp==2.6.0
anyjson==0.3.3
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.10
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.2.0
backcall==0.1.0
beautifulsoup4==4.8.1
billiard==3.6.3.0
bleach==3.1.0
bottle==0.12.16
celery==4.4.6
certifi==2020.6.20
cffi==1.12.2
chardet==3.0.4
contextlib2==0.6.0
coverage==4.5.3
cryptography==2.6.1
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-appconf==1.0.3
django-autoslug==1.9.4
django-bootstrap4==1.0.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-compressor==2.4
django-debug-toolbar==1.11
django-filter==2.2.0
django-guardian==2.3.0
django-imagekit==4.0.2
django-ipware==2.1.0
django-mdeditor==0.1.13
django-model-utils==3.2.0
django-redis==4.11.0
django-simple-rest==1.4.1
django-simpleui==2.8
django-timezone-field==4.0
django-uuslug==1.1.8
djangorestframework==3.11.0
djangorestframework-datatables==0.5.2
docopt==0.6.2
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
entrypoints==0.3
flower==0.9.5
future==0.18.2
gunicorn==20.0.0
humanize==0.5.1
idna==2.8
importlib-metadata==0.23
ipaddress==1.0.22
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.15
itsdangerous==1.1.0
jedi==0.15.1
jieba==0.39
Jinja2==2.10.3
jsonpickle==1.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
kombu==4.6.11
lazy-object-proxy==1.3.1
markdown2==2.3.7
MarkupSafe==1.1.1
mccabe==0.6.1
mimeparse==0.1.3
mistune==0.8.4
more-itertools==7.2.0
mysqlclient==1.3.13
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
olefile==0.46
packaging==19.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pilkit==2.0
Pillow==5.4.1
pipreqs==0.4.10
pluggy==0.13.0
prometheus-client==0.8.0
prompt-toolkit==2.0.10
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.3.1
pylint==2.3.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.15.4
PySocks==1.7.1
pytest==5.2.1
python-crontab==2.5.1
python-dateutil==2.8.0
python-logstash==0.4.6
python-memcached==1.59
python-slugify==3.0.0
pytz==2019.3
pyzmq==18.1.0
qtconsole==4.5.5
raven==6.10.0
rcssmin==1.0.6
redis==3.5.3
requests==2.21.0
rjsmin==1.1.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.1
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.3.1
Unidecode==1.0.23
urllib3==1.24.2
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
WeRoBot==1.8.0
whitenoise==4.1.4
Whoosh==2.7.4
widgetsnbextension==3.5.1
wrapt==1.11.1
xlrd==1.2.0
xmltodict==0.12.0
yarg==0.1.9
zipp==0.6.0
